+++ NEWS.md
@@ -5,6 +5,7 @@ explicitly setting the `registry` argument. Fixed.
 * New functions `corpus_info_file()`, `corpus_full_name()`,
 `corpus_p_attributes()`, `corpus_s_attributes()`, `corpus_properties()` and
 `corpus_property()` to retrieve registry file data.
+* New function `corpus_registry_dir()`.
 * The path to the info file in the registry file of the REUTERS corpus was
 broken. Fixed.